+++ vmm/src/lib.rs
@@ -95,6 +95,7 @@ mod serial_manager;
 #[cfg(all(feature = "kvm", feature = "sev_snp", feature = "fw_cfg"))]
 pub(crate) mod sev;
 mod sigwinch_listener;
+pub mod sparse;
 mod sync_utils;
 mod uffd;
 mod userfaultfd;
